pyproject.toml
[tool.poetry]
name = "Kitten-Claws"
version = "0.1.0"
description = "Kitties got claws...beware"
authors = ["Sir Meowsalots"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
azure-cosmos = "^4.7.0"
pytest-mock = "^3.14.0"
pytest-describe = "^2.2.0"

[tool.poetry.group.dev.dependencies]
azure-functions = "1.20.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pydantic = "^2.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
